Fixes #30216 - refactor the js session-storage
