+++ webpack/assets/javascripts/react_app/components/Layout/LayoutHelper.js
@@ -123,10 +123,3 @@ const createLocationItem = locations => {
   };
   return locItem;
 };
-
-export const checkCollapsed = () => {
-  const collapsedState = sessionStorage.getItem(
-    `["navCollapsed","pinnedPath"]`
-  );
-  return !!collapsedState && collapsedState.includes('true');
-};
