+++ webpack/assets/javascripts/react_app/components/Layout/LayoutSessionStorage.js
@@ -1,0 +1,4 @@
+import { getValue } from '../../common/SessionStorage';
+
+export const getIsNavbarCollapsed = () =>
+  !!getValue(`["navCollapsed","pinnedPath"]`);
